% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circ_emp.R
\name{Circ_emp}
\alias{Circ_emp}
\title{Running Mplus for empirical RIASEC angular locations}
\usage{
Circ_emp(
  N,
  Cor,
  konstrukt = c("R", "I", "A", "S", "E", "C"),
  showOutput = TRUE,
  Mplus_command = "Mplus",
  replaceOutfile = "always",
  name = NULL,
  ...
)
}
\arguments{
\item{N}{number of observations for correlations as numeric}

\item{Cor}{either an R matrix object with RIASEC correlations or the name of a correlation matrix, stored as a .dat text file, as character. If parameter \code{Cor} is a matrix then correlation will be saved as .dat file in current workspace directory.}

\item{konstrukt}{optionaly a character vector with length = 6 containing labels for construct dimensions - default is konstrukt = c("R","I","A","S","E","C").}

\item{showOutput}{default set to \code{TRUE} - see description of the function \code{runModels()} in package \code{\link[MplusAutomation]{MplusAutomation}}.}

\item{Mplus_command}{default set to \code{"Mplus"} as in the default in \code{\link[MplusAutomation]{MplusAutomation}} - see examples and see description of the function \code{runModels()} in package \code{\link[MplusAutomation]{MplusAutomation}}.}

\item{replaceOutfile}{default set to "always" - see description of the function \code{runModels()} in package \code{\link[MplusAutomation]{MplusAutomation}}.}

\item{name}{optional a name (as character) for the Mplus syntax to be saved - default is the object name or the name given in paramter \code{Cor} as character (with ending changed to '.inp').}

\item{...}{additional parameters passed through}
}
\value{
returns a list object containing the empirical RIASEC angular locations extracted from the Mplus result file after running the Mplus syntax.
}
\description{
This function generates and runs Mplus syntax which is discussed in Nagy et. al. (2009) and extracts the estimated model parameters - the angular locations of the six RIASEC dimensions, based on their correlations, in the asumed circumplex array within the theory of vocational choice by John Holland (1997).
}
\details{
more to come ...
}
\examples{
# generating runing and extracting parameters from Mplus files 
# refering to a correlation data stored in object ‘example2'
data(example2)# loading fictional example correlation matrix
# not Run until Mplus is installed on your Sytem #####
\dontrun{test <- Circ_emp(300,example2) 
test
### ploting the result as a circumplex
plot(test)
# for black and white printing
plot(test,ltype=c(1,2),lcolor=c("grey","grey","black","black"))}

}
\references{
Holland, J.L. (1997). \emph{Making vocational choices. A theory of vocational personalities and work environments}. Lutz, FL: Psychological Assessment Resources.

Nagy, G., Marsh, H. W., Luedtke, O., & Trautwein, U. (2009). Representing circles in our minds: Confirmatory factor analysis of circumplex structures and profiles. In T. Teo & M. S. Khine (Hrsg.), \emph{Structural Equation Modeling in Educational Research: Concepts and applications} (S. 287 - 315). Rotterdam Boston Taipei: Sense Publishers.
}
\keyword{calculus}
