% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{deg}
\alias{deg}
\title{radians to degrees}
\usage{
deg(x, m = FALSE, rev = FALSE)
}
\arguments{
\item{x}{numeric values in radians}

\item{m}{logical should values > (2 * 3.141592653589793238462643383279) be divided modulo}

\item{rev}{logical if TRUE list objekt is returnd with number of revolutions}
}
\value{
values in degrees; optionaly number of revolutions
}
\description{
This function converts angular locations from radians to degrees
}
\details{
no details
}
\examples{
## RIASEC angular locations in radians to degrees 
deg(c(0, 1.047198, 2.094395, 3.141593, 4.188790, 5.235988))
deg(6.283185)
deg(6.283185,TRUE)
deg(12.56637)
deg(12.56637,TRUE)
# 14.137167 radians is two full revolutions and a quater
# which is 90 degrees or 1.570796 radians- check it!
deg(14.137167,TRUE,TRUE)
}
\keyword{misc}
