\name{example1}
\alias{example1}
\docType{data}
\title{RIASEC correlations - perfect}
\description{
fictional correlation matrix representing a perfect RIASEC circumplex
}
\usage{
  data(example1)
}

\format{
  A 6 x 6 matrix object with named columns and rows.
}

\details{
can be written as a .dat text file using the function \code{\link{write_dat}()}.
}

\references{
Holland, J.L. (1997). \emph{Making vocational choices. A theory of vocational personalities and work environments}. Lutz, FL: Psychological Assessment Resources.
}
\examples{
data(example1)
dim(example1)
example1
}
\keyword{datasets}
