\name{example2}
\alias{example2}
\docType{data}
\title{RIASEC correlations - not so perfect}
\description{
another fictional correlation matrix representing a \emph{not so perfect} RIASEC circumplex
}
\usage{
  data(example2)
}

\format{
  A 6 x 6 matrix object with named columns and rows.
}

\details{
can be written as a .dat text file using the function \code{\link{write_dat}()}.
}

\references{
Holland, J.L. (1997). \emph{Making vocational choices. A theory of vocational personalities and work environments}. Lutz, FL: Psychological Assessment Resources.
}
\examples{
data(example2)
dim(example2)
example2
}
\keyword{datasets}
