\name{example3}
\alias{example3}
\docType{data}
\title{RIASEC and Big-Five correlations - perfect}
\description{
another fictional correlation matrix representing RIASEC (circumplex) correlations and correlations (with) five dimensions of personality (big-five). 
}
\usage{
  data(example3)
}

\format{
  A 11 x 11 matrix object with named columns and rows.
}

\details{
can be written as a .dat text file using the function \code{\link{write_dat}()}.
}

\references{
Holland, J.L. (1997). \emph{Making vocational choices. A theory of vocational personalities and work environments}. Lutz, FL: Psychological Assessment Resources.
}
\examples{
data(example3)
dim(example3)
example3
}
\keyword{datasets}
