% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.empCirc.R
\name{plot.empCirc}
\alias{plot.empCirc}
\title{S3 plotting method for object of class \code{"empCirc"}}
\usage{
\method{plot}{empCirc}(
  x,
  main = NULL,
  lcolor = c("blue", "blue", "red", "red"),
  ltype = c(1, 2),
  lwd = 3,
  defhexa = list(hexa = TRUE, seg = TRUE, gr = 5, r = 4, nseg = 7, x.cent = 0, y.cent =
    0),
  ...
)
}
\arguments{
\item{x}{object of class 'empCirc'.}

\item{main}{titel for plot.}

\item{lcolor}{(character) vector for colors; default set to \code{lcolor=c("blue","blue","red","red")}.}

\item{ltype}{vector for line types; default set to \code{ltype=c(1,2)}}

\item{lwd}{numeric defining the line width; default set to \code{lwd=3}}

\item{defhexa}{list of definition to draw the hexagon}

\item{...}{other parameters passed through}
}
\value{
a plot visualizing the empirical RIASEC angular locations within the Hexagon.
}
\description{
plotting method for object of class 'empCirc' which results from function \code{\link{Circ_emp}} and contains the empirical RIASEC angular locations extracted from the Mplus result file after running the Mplus syntax.
}
\details{
no details
}
\keyword{calculus}
