% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.proCirc.R
\name{plot.proCirc}
\alias{plot.proCirc}
\title{S3 plotting method for object of class \code{"proCirc"}}
\usage{
\method{plot}{proCirc}(
  x,
  main = NULL,
  defhexa = list(r = 4, lwd = 3, col = "grey", lty = 1, cex = 1, col.lab = "grey", gr =
    5, hexa = TRUE, seg = TRUE, nseg = 7, col.seg = "grey", x.cent = 0, y.cent = 0),
  defemp = list(r = 3, lwd = 3, col = "black", lty = 2, cex = 0.8, col.lab = "black"),
  defcon = list(r = 4.5, lwd = 3, col = "darkgray", lty = 3, cex = 1, col.lab =
    "darkgray"),
  ...
)
}
\arguments{
\item{x}{object of class 'empCirc'.}

\item{main}{titel for the plot.}

\item{defhexa}{a list with plotting options for hexagon.}

\item{defemp}{a list with plotting options for empirical RIASEC angels.}

\item{defcon}{a list with plotting options for empirical construct angels}

\item{...}{other parameters passed through}
}
\value{
a plot visualizing the empirical RIASEC and (projected) additional construct angular locations within the Hexagon.
}
\description{
plotting method for object of class 'proCirc'.
}
\details{
more to come ...
}
\keyword{calculus}
