% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad.R
\name{rad}
\alias{rad}
\title{degrees to radians}
\usage{
rad(x, m = FALSE, rev = FALSE)
}
\arguments{
\item{x}{numeric values in degrees}

\item{m}{logical should values > 360 degrees be divided modulo}

\item{rev}{logical if TRUE list objekt is returnd with number of revolutions}
}
\value{
values in radians; optionaly number of revolutions
}
\description{
This function converts angular locations from degrees to radians
}
\details{
no details
}
\examples{
## RIASEC angular locations in degrees to radians
rad(c(60,120,180,240,300,360))
rad(720)
rad(720,TRUE)
rad(360)
rad(360,TRUE)
# 810 degrees is two full revolutions and a quater
# which is 1.570796 radians or 90 degrees - check it!
rad(810,TRUE,TRUE) 



}
\keyword{misc}
