% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.deg.R
\name{rot.deg}
\alias{rot.deg}
\title{rotate degrees by amount in degrees}
\usage{
rot.deg(x, amount = 0, rev = FALSE)
}
\arguments{
\item{x}{numeric values in degrees}

\item{amount}{amount to rotate}

\item{rev}{logical if TRUE list objekt is returnd with number of revolutions}
}
\value{
values in degrees rotatet by amount; optionaly number of revolutions
}
\description{
This function rotates angular locations in degrees by an ampunt given in degrees
}
\details{
postive values vor \code{amount} will result in clockwise rotation an negative values will result in counterclockwise rotation
}
\examples{
## rotation by 30 degrees
# inital value smaler than 330 degrees
rot.deg(200,30)
rot.deg(200,30,TRUE)
# inital value smaler than 360 degrees
rot.deg(350,30)
rot.deg(350,30,TRUE)
# inital value biger than 360 degrees
rot.deg(900,30)
rot.deg(900,30,TRUE)
}
\keyword{misc}
