% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.rad.R
\name{rot.rad}
\alias{rot.rad}
\title{rotate radians by amount in radians}
\usage{
rot.rad(x, amount = 0, rev = FALSE)
}
\arguments{
\item{x}{numeric values in radians}

\item{amount}{amount to rotate (in radians)}

\item{rev}{logical if TRUE list objekt is returnd with number of revolutions}
}
\value{
values in radians rotatet by amount; optionaly number of revolutions
}
\description{
This function rotates angular locations in radians by an ampunt given in radians
}
\details{
postive values vor \code{amount} will result in clockwise rotation an negative values will result in counterclockwise rotation
}
\examples{
## rotation by 0.5235988 radians
# inital value smaler than 5.759587 radians
rot.rad(3.490659,0.5235988)
rot.rad(3.490659,0.5235988,TRUE)
# inital value smaler than 6.283185 radians
rot.rad(6.108652,0.5235988)
rot.rad(6.108652,0.5235988,TRUE)
# inital value biger than 6.283185 radians
rot.rad(15.70796,0.5235988)
rot.rad(15.70796,0.5235988,TRUE)
}
\keyword{misc}
