% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dat.R
\name{write_dat}
\alias{write_dat}
\title{writing R matrix objects as .dat text files}
\usage{
write_dat(ob, file = paste(deparse(substitute(ob)), ".dat", sep = ""), ...)
}
\arguments{
\item{ob}{the R-object to be written as .dat file.}

\item{file}{optionally the name of the .dat file as character - default is 'objectname'.dat.}

\item{...}{additional parameters passed through.}
}
\value{
a .dat text file by default written in the current workspace directory.
}
\description{
this function writes R matrix objects as .dat text files to be read by Mplus.
}
\details{
no details.
}
\examples{
## writing R-object example1 as example1.dat
\dontrun{ 
data(example1)
write_dat(example1) 
}
}
\keyword{misc}
